+++ src/js/base/ripe.js
@@ -16,7 +16,7 @@ if (
  * The version of the RIPE SDK currently in load, should
  * be in sync with the package information.
  */
-ripe.VERSION = "2.24.0";
+ripe.VERSION = "2.25.0";
 
 /**
  * Object that contains global (static) information to be used by
@@ -1191,6 +1191,27 @@ ripe.Ripe.prototype.bindImage = function(element, options = {}) {
 };
 
 /**
+ * Binds a video thumbnail Image to this Ripe instance.
+ *
+ * @param {Image} element The Image to be used by the Ripe instance.
+ * @param {Object} options An Object with options to configure the Image instance.
+ * @returns {Image} The Image instance created.
+ */
+ripe.Ripe.prototype.bindVideoThumbnail = function(element, options = {}) {
+    options = Object.assign(
+        {},
+        {
+            imageUrlProvider: this._getVideoThumbnailURL.bind(this),
+            frameValidator: this.hasVideo.bind(this),
+            doubleBuffering: false
+        },
+        options
+    );
+    const image = new ripe.Image(this, element, options);
+    return this.bindInteractable(image);
+};
+
+/**
  * Binds an Configurator to this Ripe instance.
  *
  * @param {Configurator} element The Configurator to be used by the Ripe instance.
@@ -1632,6 +1653,7 @@ ripe.Ripe.prototype.getDimension = function(name = "$base", face = null) {
     }
 
     if (!face) return dimensionSpec;
+    if (!dimensionSpec.faces) return dimensionSpec;
 
     return dimensionSpec.faces[face] ? dimensionSpec.faces[face] : dimensionSpec;
 };